Migrate methods into Processor classes
